backend/requirements.txt
aniso8601==7.0.0
astroid==2.2.5
beautifulsoup4==4.7.1
blis==0.2.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cobble==0.1.3
cymem==2.0.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
mammoth==1.4.9
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.2
numpy==1.16.4
plac==0.9.6
preshed==2.0.1
pylint==2.3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
spacy==2.1.6
srsly==0.0.7
thinc==7.0.8
tqdm==4.32.2
typed-ast==1.4.0
urllib3==1.25.3
uWSGI==2.0.18
wasabi==0.2.2
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
Pillow==6.1.0
